Implement CAEN-V1742
A tool reading out the CAEN V1742 digitizer.

See test/read-v1742.cpp for an example of readout.

Currently the reading stops after 1 event. Delete line 264 in
UserTools/CAEN-V1742/CAEN-V1742.cpp to read more. You will need to
figure out how to properly stop the acquisition then since if Finalise()
is not called, the ROOT file is not closed properly.

The board produces waveforms as arrays of float, but ROOT doesn't know
how to store std::vector<float> out of the box, so this tool converts
them to std::vector<double>. Should we create a ROOT dictionary for
std::vector<float> or agree on the way of storing raw C arrays, this
would speed up the code and reduce memory and disk usage.