ThreadLoop: revise interface
ThreadLoop::subscribe used to return an iterator to a private list as a
handle. Before C++14 there is no way to tell whether such an iterator is
initialized, so tools had to store a pointer to this iterator and
nullify the pointer when the thread is terminated. Now ThreadLoop::add
(which ThreadLoop::subscribe was renamed to) returns an object inspired
by std::thread. This object holds a reference to the ThreadLoop it was
created with and use it to control its own execution, in particular
termination. Other thread controls similar to that in POSIX API can now
be transparently implemented if desired.

Rename
ThreadLoop::subscribe -> ThreadLoop::add

Delete
ThreadLoop::unsubscribe (use Thread::terminate instead)
ThreadLoop::pause  (not used, can be easily reimplemented if needed)
ThreadLoop::resume (ditto)
ThreadLoop::paused (ditto)
ThreadLoop::clear  (ditto)