V812, V1495: read register settings from the database
For V1495 (FPGA), if the configuration JSON has an entry "config", treat
it as an object with keys being register addresses and values being the
values that should be written to the register addresses.

V812 (CFD) configuration is similar, except that we support a single
V1495 but multiple V812s, so for V812 there is an extra level of
indirection: the "config" object should map CFD VME addresses to
registers settings.